+++ examples/country.py
@@ -33,15 +33,16 @@ class Country(Place):
     cities: List[City]
 
 
-current_module = sys.modules["__main__"]
-mg = MermaidGenerator(current_module)
+if __name__ == "__main__":
+    current_module = sys.modules["__main__"]
+    mg = MermaidGenerator(current_module)
 
-chart_dependency = mg.generate_chart(root="Country")
-with open("./examples/country.md", mode="w") as f:
-    f.write(chart_dependency)
-    f.close()
+    chart_dependency = mg.generate_chart(root="Country")
+    with open("./examples/country.md", mode="w") as f:
+        f.write(chart_dependency)
+        f.close()
 
-chart_all = mg.generate_chart(relations=Relations.Dependency | Relations.Inheritance)
-with open("./examples/country_all.md", mode="w") as f:
-    f.write(chart_all)
-    f.close()
+    chart_all = mg.generate_chart(relations=Relations.Dependency | Relations.Inheritance)
+    with open("./examples/country_all.md", mode="w") as f:
+        f.write(chart_all)
+        f.close()