+++ pydantic_mermaid/__init__.py
@@ -1,1 +1,0 @@
-from pydantic_mermaid.mermaid_generator import MermaidGenerator, Relations  # noqa