+++ src/pydantic_mermaid/__init__.py
@@ -1,0 +1,3 @@
+from pydantic_mermaid.__main__ import main  # noqa F401
+from pydantic_mermaid.mermaid_generator import MermaidGenerator  # noqa F401
+from pydantic_mermaid.models import *  # noqa F401