+++ src/pydantic_mermaid/__main__.py
@@ -8,6 +8,7 @@ from types import ModuleType
 from uuid import uuid4
 
 from pydantic_mermaid.mermaid_generator import MermaidGenerator
+from pydantic_mermaid.models import Relations
 
 logger = logging.getLogger("pydantic-mermaid")
 
@@ -29,7 +30,7 @@ def import_module(path: str) -> ModuleType:
             return module
         else:
             return importlib.import_module(path)
-    except Exception as e:
+    except Exception as e:  # pragma: no cover
         logger.error("The --module argument must be a module path separated by dots or a valid filepath")
         raise e
 
@@ -44,14 +45,32 @@ def _parse_cli_args() -> argparse.Namespace:
         formatter_class=argparse.RawTextHelpFormatter,
     )
     parser.add_argument(
+        "-m",
         "--module",
+        type=str,
         help="name or filepath of the python module.\n" "Discoverable submodules will also be checked.",
     )
     parser.add_argument(
+        "-o",
         "--output",
+        type=str,
         help="name of the file the mermaid chart should be written to.",
     )
-    parser.add_argument("--root", help="Root class for dependency chart or inheritence chart", default="")
+    parser.add_argument(
+        "-n",
+        "--root",
+        type=str,
+        help="Root node for dependency chart or inheritance chart",
+        default="",
+    )
+    parser.add_argument(
+        "-e",
+        "--relations",
+        nargs="+",
+        type=str,
+        help="Dependency or Inheritance chart",
+        default="dependency",
+    )
     return parser.parse_args()
 
 
@@ -63,10 +82,18 @@ def main():
     module_type = import_module(args.module)
 
     mg = MermaidGenerator(module_type)
-    s = mg.generate_chart(root=args.root)
+    relations = Relations.Dependency
+    if "dependency" in args.relations and "inheritance" in args.relations:
+        relations = Relations.Dependency | Relations.Inheritance
+    elif "dependency" in args.relations:
+        relations = Relations.Dependency
+    elif "inheritance" in args.relations:
+        relations = Relations.Inheritance
+
+    chart_content = mg.generate_chart(root=args.root, relations=relations)
     with open(args.output, "w") as f:
-        f.write(s)
+        f.write(chart_content)
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
     main()